% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sim.R
\name{augment_mcmc}
\alias{augment_mcmc}
\title{Augment MCMC}
\usage{
augment_mcmc(object, dat, ..., verbose = TRUE, force = FALSE)
}
\arguments{
\item{object}{(Gaussian) output of BiDAG MCMC}

\item{dat}{data frame used in sampling}

\item{...}{other arguments}

\item{verbose}{logical indicating whether to show progress}

\item{force}{logical - can this be used on previously augmented chains?}
}
\description{
Adds samples and causal effects to partition MCMC output
for Gaussian data.
}
