% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sim.R
\name{causalEffects}
\alias{causalEffects}
\title{Get matrix of marginal causal effects for every pair of variables}
\usage{
causalEffects(A, Sigma)
}
\arguments{
\item{A}{adjacency matrix of DAG}

\item{Sigma}{covariance matrix (assumed Markov with respect to \code{A})}
}
\value{
gives matrix whose i,j th entry is total
effect of i on j.
}
\description{
Get matrix of marginal causal effects for every pair of variables
}
\details{
Gaussian data only.
}
