% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_out.R
\name{fit_mcmc}
\alias{fit_mcmc}
\title{Set up MCMC Chain}
\usage{
fit_mcmc(data, scoretype = "bge", iterations, ...)
}
\arguments{
\item{data}{dataset to fit}

\item{scoretype}{type of score: "bge" or "bde"}

\item{iterations}{number of iterations}

\item{...}{other arguments to pass to \code{partitionMCMC}}
}
\description{
Set up MCMC Chain
}
\details{
This is basically a wrapper for
BiDAG's \code{scoreparameters()} and \code{partitionMCMC()},
but which augments
the chain with the variable names extracted
from the original data set.  Returns an object of class
\code{MCMCchain}.
}
