% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_out.R
\name{fit_multiple}
\alias{fit_multiple}
\title{Set up MCMC Chain}
\usage{
fit_multiple(data_sets, scoretype = "bge", iterations, ...)
}
\arguments{
\item{data_sets}{list of datasets}

\item{scoretype}{type of score: "bge" or "bde"}

\item{iterations}{number of iterations}

\item{...}{other arguments to pass to \code{partitionMCMC}}
}
\description{
Set up MCMC Chain
}
\details{
This will run \code{fit_mcmc} on multiple chains,
augment the output for each, and then join the outputs
together.  Intended for cases where we have multiple
imputations of the dataset.
}
