% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funcs.R
\name{gaussianize}
\alias{gaussianize}
\title{Gaussianize Data}
\usage{
gaussianize(data, jitter = FALSE)
}
\arguments{
\item{data}{data set (variables as columns)}

\item{jitter}{logical: if \code{TRUE}, ties are broken by jittering}
}
\description{
Gaussianize Data
}
