% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sim.R
\name{genSamp}
\alias{genSamp}
\title{Generate sample from posterior distribution of (Gaussian) BN}
\usage{
genSamp(A, xbar, S, n, alpha, T0)
}
\arguments{
\item{A}{adjacency matrix for graph}

\item{xbar}{sample mean of data}

\item{S}{sample variance of data (not normalized by \code{n})}

\item{n}{number of samples}

\item{alpha}{prior parameter, default to dimension of data}

\item{T0}{prior precision, defaults to identity matrix}
}
\description{
Generate sample from posterior distribution of (Gaussian) BN
}
