% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sim.R
\name{intervene}
\alias{intervene}
\title{Fix Gaussian distribution by dividing by p(a | b)}
\usage{
intervene(Sigma, a, b, cov_only = TRUE)
}
\arguments{
\item{Sigma}{covariance matrix}

\item{a, b}{sets}

\item{cov_only}{logical (currently \code{TRUE} is only acceptable value)

Returns covariance matrix of same dimensions as Sigma, with
dependence of variables in \code{a} on variables in \code{b}
removed.}
}
\description{
Fix Gaussian distribution by dividing by p(a | b)
}
