% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sim.R
\name{merge_chain}
\alias{merge_chain}
\title{Merge two MCMC chains into one}
\usage{
merge_chain(object, ...)
}
\arguments{
\item{object}{chain or list of chains}

\item{...}{other chains

Combine separate chains into a single sample.  Should
work for \code{MCMCchain}, \code{aug_chain} or \code{summary_MCMCchain}.}
}
\description{
Merge two MCMC chains into one
}
