% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediators.R
\name{my_sum}
\alias{my_sum}
\title{Standard summary to use}
\usage{
my_sum(x, alpha = 0.1)
}
\arguments{
\item{x}{numerical vector}

\item{alpha}{one minus the coverage}
}
\description{
Standard summary to use
}
