% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funcs.R
\name{params}
\alias{params}
\title{Identify SEM parameters for a DAG}
\usage{
params(Sigma, graph)
}
\arguments{
\item{Sigma}{covariance matrix (either in topological order or with \code{graph} specified)}

\item{graph}{optionally, adjacency matrix object for underlying graph}
}
\value{
list of \code{B} (upper triangular) and \code{Om} (diagonal)
such that if IB = (I-B)^{-1} then \code{t(IB) \%*\% Om \%*\% IB} gives \code{Sigma}.
}
\description{
Identify SEM parameters for a DAG
}
\details{
The order of variables is assumed toplogical if the graph
is not specified.
}
