% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_out.R
\name{plot.summary_MCMCchain}
\alias{plot.summary_MCMCchain}
\title{Plot average adjacency from MCMC output}
\usage{
\method{plot}{summary_MCMCchain}(x, p_thresh = 0.5, ...)
}
\arguments{
\item{x}{object of class \code{summary.MCMCchain}}

\item{p_thresh}{threshold for plotting an edge (default is 0.5)}

\item{...}{other arguments}
}
\description{
Plot average adjacency from MCMC output
}
\details{
This plots an undirected and directed summary
of the output.  The method will only plot an edge if it appears
in at least a proportion \code{p_thresh} of the samples.
Consequently, an undirected edge may appear but no
corresponding directed edge.
}
