% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_causal_effects}
\alias{plot_causal_effects}
\title{Plot grid summarizing causal effects}
\usage{
plot_causal_effects(x, subset, alpha = 0.05)
}
\arguments{
\item{x}{object of class \code{aug_chain}}

\item{subset}{subset of variables to use}

\item{alpha}{significance level}
}
\description{
Plot grid summarizing causal effects
}
\details{
Given the output of \code{augment_chain},
this plots pairwise causal effect histograms in
a grid.
}
