% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot graph from adjacency matrix}
\usage{
plot_graph(amat, p_thresh = 0.5, edgemode = "directed")
}
\arguments{
\item{amat}{adjacency matrix}

\item{p_thresh}{minimum value to plot}

\item{edgemode}{either \code{"directed"} or \code{"undirected"}}
}
\description{
Plot graph from adjacency matrix
}
