% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_measures.R
\name{sprmn}
\alias{sprmn}
\title{Implement Spearmen's rank for ordinal data}
\usage{
sprmn(x, ..., use = "pairwise")
}
\arguments{
\item{x}{a numeric or ordinal vector}

\item{...}{further numeric or ordinal vectors}

\item{use}{data to use}
}
\description{
Implement Spearmen's rank for ordinal data
}
