% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediators.R
\name{sum_causal_paths}
\alias{sum_causal_paths}
\title{Summarize causal relationship between two variables}
\usage{
sum_causal_paths(
  object,
  i,
  j,
  digits = 2,
  alpha = 0.1,
  show = TRUE,
  indirect = TRUE
)
}
\arguments{
\item{object}{augmented output of MCMC algorithm}

\item{i, j}{variables whose relationship is to be analysed}

\item{digits}{how to round numbers in printed output (see details)}

\item{alpha}{mass to remove from confidence interval (so defaults to 90\% CI)}

\item{show}{logical: should ACE be printed?}

\item{indirect}{logical: if true includes indirect effects}
}
\description{
Summarize causal relationship between two variables
}
\details{
Get summary of causal connections and mediations.
Prints summary and returns list of \code{total_eff},
\code{direct_eff} and \code{indirect_eff}, each being a list
with entries \code{to} and \code{from}.

The output is printed as a mixture of percentages and
causal effects, each of which is rounded to 2 significant
figures by default.  This value can be modified by
using the \code{digits} argument.
}
