% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_out.R
\name{summary.MCMCtrace}
\alias{summary.MCMCtrace}
\title{Summarise output from MCMC}
\usage{
\method{summary}{MCMCtrace}(object, ...)
}
\arguments{
\item{object}{output from \code{BiDAG} package}

\item{...}{other arguments}
}
\description{
Summarise output from MCMC
}
\details{
Constructs the average adjacency matrix
and mean sparsity level.  Replaces list of DAG and partition
scores with vectors.  Also has a print and plot
method.
}
