% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikz.R
\name{TikzCode}
\alias{TikzCode}
\title{Generate Tikz Code for Plot}
\usage{
TikzCode(
  x,
  file = "",
  z,
  n_pos,
  bend,
  ud_thresh = 0.5,
  cutoff = 0.1,
  func,
  bend_double = TRUE,
  rev = FALSE,
  col = c("black", "black", "blue", "red"),
  labels = NULL,
  nodesOnTop = TRUE,
  edgeScale = 1,
  ...
)
}
\arguments{
\item{x}{adjacency matrix from MCMC summary}

\item{file}{file to output code (defaults to stout)}

\item{z}{optionally, a matrix of z-scores}

\item{n_pos}{optional matrix of node positions (requires Rgraphviz if not specified)}

\item{bend}{edge angles}

\item{ud_thresh}{threshold for drawing edges as undirected}

\item{cutoff}{threshold for including edges}

\item{func}{vectorized function to transform probabilities}

\item{bend_double}{bend automatically if arrows in two different directions?}

\item{rev}{logical - should thickness and density plotting be reversed?}

\item{col}{colours for undirected, directed and positive and negative edges}

\item{labels}{character vector of labels for nodes}

\item{nodesOnTop}{logical: should nodes be on top of everything else}

\item{edgeScale}{scales edges to a desired width}

\item{...}{other arguments}
}
\description{
Generate Tikz Code for Plot
}
\details{
If more than \code{ud_thresh} of the edges present
are in a particular direction, then the two directions are drawn
separately, otherwise an undirected edge is drawn.  Edges present
in less than \code{cutoff} sampled graphs are ignored entirely. \code{func}
is a transformation that can be applied to the probability (e.g.
squaring it to reduce the visual impact of weak edges).  This must
take a matrix and return a matrix of the transformed values.

If \code{z} is specified, then the function will plot the thickness
of the line to be related to this quantity.
If \code{rev = TRUE} then the thickness of the line is related to the
size of the effect, and the density of colour to the proportion of
samples containing that effect.

\code{col} is given as a character vector of up to four colours.
The order is undirected positive and negative, followed by
directed positive and negative.  Colours are just repeated if
necessary.

If \code{labels} is not provided these are taken from the
rownames of \code{x}.
}
